Add files via upload
Initially some of the letters are lowercase, while others are uppercase(i.e. capitals).The aim of the game is to get all the letters to be capitals. In order to change the letters in the grid, the user will select a letter then the selected letter and all adjacent letters in the same row or column (e.g. ‘l’, ‘o’, ‘p’) are toggled (uppercase to lowercase or lowercase
to uppercase).This continues until the player succeeds or decides to quit. Each time the player plays a new game, randomly decide which letters will be uppercase or lowercase initially.
The selecting of a letter should trigger the letter itself and adjacent letters to toggle. Toggling values can trigger the calculation of whether the game has been won or not.
About the game :
In object oriented programming,This type of behavioural problem and the solution is so common in software development that it has
a name, specifically the Observer Pattern. A design pattern is defined as a general reusable solution to a commonly occurring problem and there
are many such patterns. This pattern is also the basis of many event-driven APIs. 
The capitals game implements the observer pattern by attaching listeners to cells when a users wants to change from small letter to capital letters.